// src/enums/ItemTypes.ts
export enum WeaponTypes {
  WEAPONS = "Weapons",
  MELEE_WEAPONS = "Melee Weapons",
  AXES = "Axes",
  MACES = "Maces",
  SWORDS = "Swords",
  STAVES = "Staves",
  CLUBS = "Clubs",
  POLEARMS = "Polearms",
  SCEPTERS = "Scepters",
  HAMMERS = "Hammers",
  MISSILE_WEAPONS = "Missile Weapons",
  CROSSBOWS = "Crossbows",
  BOWS = "Bows",
  JAVELINS = "Javelins",
  DAGGERS = "Daggers",
  SPEARS = "Spears",
  THROWING = "Throwing",
  WANDS = "Wands",
  AMAZON_WEAPONS = "Amazon Weapons",
  SORCERESS_ORBS = "Sorceress Orbs",
  ASSASSIN_KATARS = "Assassin Katars"
}

export enum ArmorTypes {
  ARMOR = "Armor",
  SHIELDS = "Shields",
  HELMS = "Helms",
  PALADIN_SHIELDS = "Paladin Shields",
  NECROMANCER_SHRUNKEN_HEADS = "Necromancer Shrunken Heads",
  BARBARIAN_HELMS = "Barbarian Helms",
  DRUID_PELTS = "Druid Pelts"
}

export enum Slots {
  WEAPON,
  ARMOR,
  HELM,
  SHIELD,
}